Feat: Coin History 리포트 추적

+++ src/main/java/com/knj/mirou/boundedContext/member/model/dtos/CoinReportDTO.java
@@ -13,8 +13,8 @@ public class CoinReportDTO {
     public int weeklyGivenCoinSum;
     public int weeklyUsedCoinSum;
 
-    public int weeklyGivenCoinAverage;
-    public int weeklyUsedCoinAverage;
+    public long weeklyGivenCoinAverage;
+    public long weeklyUsedCoinAverage;
 
     public int topCoinUserAmount;
 }
